+++ SBC Admin/src/pages/trackList/TrackList.jsx
@@ -4,6 +4,7 @@ import { Link } from 'react-router-dom';
 import { IconButton } from '@mui/material';
 import DeleteOutline from '@mui/icons-material/DeleteOutline';
 import "./tracklist.css";
+import Swal from 'sweetalert2';
 
 export default function TrackList() {
   const [data, setData] = useState([]);
@@ -16,17 +17,34 @@ export default function TrackList() {
   }, []);
 
   const handleDelete = (id) => {
-    fetch(`/musictest/track/delete/${id}`, {
-      method: 'DELETE',
-    })
-      .then((response) => {
-        if (response.ok) {
-          setData(data.filter((item) => item.trackId !== id));
-        } else {
-          console.error("Error deleting track:", response.statusText);
-        }
-      })
-      .catch((error) => console.error("Error deleting track:", error));
+    Swal.fire({
+      title: "Are you sure you want to delete this track?",
+      showDenyButton: true,
+      showCancelButton: true,
+      confirmButtonText: "Delete",
+      denyButtonText: `Don't delete`
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`/musictest/track/delete/${id}`, {
+          method: 'DELETE',
+        })
+          .then((response) => {
+            if (response.ok) {
+              setData(data.filter((item) => item.trackId !== id));
+              Swal.fire("Deleted!", "Track has been deleted.", "success");
+            } else {
+              console.error(`Error deleting track: ${response.statusText}`);
+              Swal.fire("Error!", `Error deleting track: ${response.statusText}`, "error");
+            }
+          })
+          .catch((error) => {
+            console.error("Error deleting track:", error);
+            Swal.fire("Error!", "Error deleting track.", "error");
+          });
+      } else if (result.isDenied) {
+        Swal.fire("Cancelled", "Track was not deleted", "info");
+      }
+    });
   };
 
   const columns = [
